ui/engines/requirements.txt
pillow>=9.0.0
diffusers>=0.25.0
torch>=2.0.0
numpy>=1.20.0 